add filename correction 2.0
